refactor(game-engine): extract handleBallPickup + canUseTeamReroll (S27.8.5)
Slice 5 de S27.8 (refactor monolith `actions.ts`).

Extraction (`actions/ball-pickup.ts`)
- `handleBallPickup` : ramassage de balle complet (skill `no-hands`,
  modifiers basics + skills, relance Sure Hands, succes/touchdown,
  echec / team reroll / turnover + rebond).
- Aucune dependance interne au dispatcher : tout passe par les
  modules `mechanics/*`, `skills/*`, `utils/*` et `core/game-state`.

Helper deplace (`core/game-state.ts`)
- `canUseTeamReroll(state, team)` migre depuis `actions.ts`. Le
  fichier est la place naturelle pour ce check pur sur `state`
  (a cote de `canPlayerMove` / `canPlayerContinueMoving` /
  `getGfiCap`). Permet a `ball-pickup.ts` (et aux extractions
  ulterieures) de l'importer sans creer de cycle vers
  `actions.ts`.

Resultat
- `actions.ts` 2240 -> 2135 lignes (-105 nettes ; cumul -571 vs
  origine 2706 = ~21.1%).
- `ball-pickup.ts` ajoute a la liste d'exclusion coverage.
- 4888 tests game-engine passent. Coverage 25.33% maintenue.

Reste pour atteindre la cible <= 600 lignes :
- handleLeap (depend de getLegalMoves, cycle a casser)
- handleNormalMove / handleDodgeRoll / handleDodge / handleMove
- handleBlock / choice handlers
- handleBlitz
